requirements.txt
numpy>=1.19.2
scipy>=1.6.3
librosa~=0.9.1
SoundFile~=0.10
resampy==0.2.2
numba>=0.53.0
pydub~=0.25.1
ffmpeg-python>=0.2.0
matplotlib>=3.5.2
termcolor>=1.1.0
pyaudio>=0.2.11
typeguard>=2.13.3
pillow>=8.3.2
tqdm>=4.64.1
visualdl>=2.2.3
pyyaml>=5.4.1
scikit-learn>=1.0.2